feat(cb2-3842): merge isssues
