+++ src/app/features/test-records/views/amended-test-record/amended-test-record.component.ts
@@ -4,6 +4,7 @@ import { DefectTpl } from '@forms/templates/general/defect.template';
 import { masterTpl } from '@forms/templates/test-records/master.template';
 import { Defects } from '@models/defects';
 import { TestResultModel } from '@models/test-result.model';
+import { VehicleTypes } from '@models/vehicle-tech-record.model';
 import { TestRecordsService } from '@services/test-records/test-records.service';
 import { Observable } from 'rxjs';
 
@@ -13,14 +14,20 @@ import { Observable } from 'rxjs';
 })
 export class AmendedTestRecordComponent {
   testResult$: Observable<TestResultModel | undefined>;
-  template: FormNode;
   defectTpl: FormNode;
   defectData$: Observable<Defects | undefined>;
 
   constructor(private testRecordsService: TestRecordsService) {
     this.testResult$ = this.testRecordsService.amendedTestResult$;
-    this.template = PsvAnnual;
     this.defectTpl = DefectTpl;
     this.defectData$ = this.testRecordsService.amendedDefectData$;
+  }
+
+  get vehicleTypes() {
+    return VehicleTypes;
+  }
+
+  get masterTpl() {
+    return masterTpl;
   }
 }
